requirements.txt
flask
sqlalchemy
psycopg2-binary
Flask-SQLAlchemy
Werkzeug